feat : 포스트 추천 기능

+++ backend/routers/post.py
@@ -11,6 +11,7 @@ from domain.post_crud import (
     get_post_list,
     update_post,
     delete_post,
+    vote_post,
 )
 from .auth import get_current_user
 
@@ -20,6 +21,10 @@ user_dependency = Annotated[dict, Depends(get_current_user)]
 
 
 class DeletePostRequest(BaseModel):
+    post_id: int
+
+
+class VotePostRequest(BaseModel):
     post_id: int
 
 
@@ -35,7 +40,7 @@ def post_detail_api(post_id: int):
     return post
 
 
-@router.post("/create")
+@router.post("/create", status_code=status.HTTP_201_CREATED)
 def create_post_api(post_request: PostRequest, user: user_dependency):
     try:
         user_id = user.get("id")
@@ -46,15 +51,34 @@ def create_post_api(post_request: PostRequest, user: user_dependency):
 
 
 @router.put("/update/{post_id}")
-def update_post_api(post_request: PostRequest, post_id: int):
-
-    update_post(post_id, post_request)
+def update_post_api(post_request: PostRequest, post_id: int, user: user_dependency):
+    try:
+        user_id = user.get("id")
+        post = get_post(post_id)
+        if post["author_id"] != user_id:
+            raise HTTPException(status_code=401, detail="Not Authorized to update post")
+        update_post(post_id, post_request)
+    except Exception as e:
+        raise HTTPException(status_code=400, detail=f"Failed to update post: {str(e)}")
 
     return {"detail": "Post updated successfully"}
 
 
 @router.delete("/delete/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_post_api(post_id: int):
-    delete_post(post_id)
+def delete_post_api(post_id: int, user: user_dependency):
+    try:
+        user_id = user.get("id")
+        post = get_post(post_id)
+        if post["author_id"] != user_id:
+            raise HTTPException(status_code=401, detail="Not Authorized to delete post")
+        delete_post(post_id)
+    except Exception as e:
+        raise HTTPException(status_code=400, detail=f"Failed to update post: {str(e)}")
 
     return {"detail": "Post deleted successfully"}
+
+
+@router.post("/vote")
+def post_vote_api(request: VotePostRequest, user: user_dependency):
+    user_id = user.get("id")
+    vote_post(user_id, request.post_id)
